feat: Add hook and context for React
