+++ packages/react-hook/README.md
@@ -15,29 +15,25 @@ npm install @near-wallet-selector/react-hook
 Then use it in your dApp:
 
 ```ts
+import '@near-wallet-selector/modal-ui/styles.css';
 import { setupMyNearWallet } from '@near-wallet-selector/my-near-wallet';
 import { setupMeteorWallet } from '@near-wallet-selector/meteor-wallet';
-import { useWalletSelector } from '@near-wallet-selector/react'
+import { WalletSelectorProvider } from '@near-wallet-selector/react-hook';
+ 
+const walletSelectorConfig = {
+  network: "testnet",
+  createAccessKeyFor: "hello.near-examples.testnet",
+  modules: [
+    setupMyNearWallet(),
+    setupMeteorWallet(),
+  ],
+}
 
-export default function MyApp({ Component, pageProps }) {
-
-  const { setupWalletSelector } = useWalletSelector();
-
-  useEffect(() => {
-    setupWalletSelector({
-      network: NetworkId,
-      createAccessKeyFor: HelloNearContract,
-      modules: [
-        setupMyNearWallet(),
-        setupMeteorWallet(),
-      ],
-    })
-  }, []);
-
-  return (<>
-      <Navigation />
+export default function App({ Component }) {
+  return (
+    <WalletSelectorProvider config={walletSelectorConfig}>
       <Component {...pageProps} />
-    </>
+    </WalletSelectorProvider>
   );
 }
 ```
@@ -46,8 +42,12 @@ export default function MyApp({ Component, pageProps }) {
 import { useWalletSelector } from '@near-wallet-selector/react'
 
 export default function OtherComponent() {
-  const { signedAccountId, walletSelector } = useWalletSelector();
+  const { signedAccountId, walletSelector, wallet, modal } = useWalletSelector();
   ...
+  // modal.show() to show the wallet selector modal
+  // walletSelector.signOut() to log out the user
+  ...
+  }
 ```
 
 ## Options
