+++ packages/react-hook/src/lib/useWalletSelector.tsx
@@ -1,10 +1,10 @@
 import { useContext } from "react";
-import { WalletSelectorContext } from "../context/WalletSelectorProvider";
+import { WalletSelectorContext } from "./WalletSelectorProvider";
 
 export function useWalletSelector() {
   const context = useContext(WalletSelectorContext);
   if (!context) {
-    throw new Error("useWalletSelector must be used within <WalletSelectorProvider>");
+    throw new Error("useWalletSelector must be used within a <WalletSelectorProvider>");
   }
   return context;
 }
